refactor(collab): split forkSession into relay/workspace/handoff
Move the durable fork-relay registry, the workspace resolver plus setup
and checkout dialogs, and the bounded first-send handoff digest into
flat siblings; forkSession.ts keeps forkTeammateSession and re-exports
everything, so all 23 import sites are unchanged. The localStorage
registry key now has exactly one owner. No behaviour change.
